pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tmux-restore"]

[project]
name = "tmux-restore"
version = "0.0.2"
authors = [
    { name = "Arjun Ghoshal", email = "arjunp.ghoshal@gmail.com" }
]
description = "A python executable that saves and restores tmux sessions across reboots"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "libtmux",
    "pyyaml",
    "dacite",
]

[project.urls]
Homepage = "https://github.com/arjunghoshal/tmux-restore"
Issues = "https://github.com/arjunghoshal/tmux-restore/issues"

[project.optional-dependencies]
test = [
    "pytest"
]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
log_cli = 1
log_cli_level = "INFO"
log_format = "%(levelname)s: %(message)s"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not integration\"')"
]
